[finelog] Drop LogClient.push_batch with the relay that needed it
`push_batch` arrived with the Python relay (#6871) as its durable-ack write
and never gained another caller: the worker's log path uses `write_batch`,
and iris tests already push through the generated `LogServiceClientSync`.
Point the one remaining caller at that client and delete the method.

The `log_service` fixture it needs was defined identically in two conftests;
hoist one copy to the iris test root.
