+++ lib/iris/tests/cluster/test_attempt_logs.py
@@ -7,8 +7,8 @@ Verifies that task status correctly records multiple attempts after
 failure + retry, using the ServiceTestHarness (parameterized GCP + K8s).
 """
 
-from finelog.client import LogClient
 from finelog.rpc import logging_pb2
+from finelog.rpc.logging_connect import LogServiceClientSync
 from iris.cluster.log_keys import task_log_key
 from iris.cluster.types import TaskAttempt
 from iris.rpc import controller_pb2, job_pb2
@@ -16,14 +16,15 @@ from iris.rpc import controller_pb2, job_pb2
 from .conftest import ServiceTestHarness
 
 
-def _push_task_logs(log_client: LogClient, task_id, lines: list[str]) -> None:
+def _push_task_logs(log_service: LogServiceClientSync, task_id, lines: list[str]) -> None:
     """Durably push ``lines`` as the first attempt's logs for ``task_id``."""
     entries = []
     for i, data in enumerate(lines):
         entry = logging_pb2.LogEntry(source="stdout", data=data)
         entry.timestamp.epoch_ms = 1000 + i
         entries.append(entry)
-    log_client.push_batch(task_log_key(TaskAttempt(task_id=task_id, attempt_id=0)), entries)
+    key = task_log_key(TaskAttempt(task_id=task_id, attempt_id=0))
+    log_service.push_logs(logging_pb2.PushLogsRequest(key=key, entries=entries))
 
 
 def test_task_status_shows_attempts(harness: ServiceTestHarness):
@@ -59,7 +60,7 @@ def test_task_status_shows_attempts(harness: ServiceTestHarness):
     assert attempts[-1].state == job_pb2.TASK_STATE_SUCCEEDED
 
 
-def test_get_task_status_surfaces_root_cause_highlights(harness: ServiceTestHarness, log_client: LogClient):
+def test_get_task_status_surfaces_root_cause_highlights(harness: ServiceTestHarness, log_service: LogServiceClientSync):
     """A failed task's status distills the crash out of its noisy logs."""
     if harness.provider_type == "gcp":
         harness.register_gcp_worker("w1")
@@ -69,7 +70,7 @@ def test_get_task_status_surfaces_root_cause_highlights(harness: ServiceTestHarn
     harness.drive_task_state(task_id, job_pb2.TASK_STATE_FAILED)
 
     _push_task_logs(
-        log_client,
+        log_service,
         task_id,
         [
             " 50%|#####     | 500/1000 [00:10<00:10,  5.0it/s]",
@@ -88,7 +89,9 @@ def test_get_task_status_surfaces_root_cause_highlights(harness: ServiceTestHarn
     assert not any("500/1000" in line for line in resp.root_cause_highlights)
 
 
-def test_get_task_status_no_highlights_for_succeeded_task(harness: ServiceTestHarness, log_client: LogClient):
+def test_get_task_status_no_highlights_for_succeeded_task(
+    harness: ServiceTestHarness, log_service: LogServiceClientSync
+):
     """A task that did not fail gets no root-cause highlights, even with logs."""
     if harness.provider_type == "gcp":
         harness.register_gcp_worker("w1")
@@ -97,7 +100,7 @@ def test_get_task_status_no_highlights_for_succeeded_task(harness: ServiceTestHa
     task_id = harness._query_tasks(job_id)[0].task_id
     harness.drive_task_state(task_id, job_pb2.TASK_STATE_SUCCEEDED)
 
-    _push_task_logs(log_client, task_id, ["RuntimeError: this ran on an earlier, since-retried attempt"])
+    _push_task_logs(log_service, task_id, ["RuntimeError: this ran on an earlier, since-retried attempt"])
 
     resp = harness.service.get_task_status(
         controller_pb2.Controller.GetTaskStatusRequest(task_id=task_id.to_wire()), None
