+++ lib/iris/tests/conftest.py
@@ -17,6 +17,7 @@ import pytest
 from finelog.client import LogClient
 from finelog.embedded import is_available as finelog_native_available
 from finelog.embedded import require_embedded_server
+from finelog.rpc.logging_connect import LogServiceClientSync
 from iris.client.local_client import make_local_client
 from iris.cluster.config import (
     IrisClusterConfig,
@@ -66,6 +67,18 @@ def log_client(embedded_log_server):
         yield client
     finally:
         client.close()
+
+
+@pytest.fixture
+def log_service(embedded_log_server) -> LogServiceClientSync:
+    """A LogService RPC client against the per-test embedded server.
+
+    The native server makes pushed log entries immediately fetchable (RAM
+    buffer), so push→fetch is synchronously visible within a test without any
+    manual flush. The sync client exposes ``push_logs(request)`` /
+    ``fetch_logs(request)``.
+    """
+    return LogServiceClientSync(address=embedded_log_server.address)
 
 
 def _make_controller_only_config() -> IrisClusterConfig:
